ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Serwis AI obsługujący operacje LangChain i integrację z LLM dla projektu TEG"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.11.0",
    "flask>=3.1.1",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.22",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
    "supabase>=2.15.3",
    "pyyaml>=6.0",
    "requests>=2.32.3",
]
